refactor: 패키지명 place -> restaurant 변경
