+++ backend/hankkimoa/src/main/java/com/jjangiji/hankkimoa/expense/service/ExpenseService.java
@@ -8,8 +8,8 @@ import com.jjangiji.hankkimoa.expense.repository.ExpenseEntity;
 import com.jjangiji.hankkimoa.expense.repository.ExpenseRepository;
 import com.jjangiji.hankkimoa.expense.repository.ExpenseSavingGoalEntity;
 import com.jjangiji.hankkimoa.expense.repository.ExpenseSavingGoalRepository;
-import com.jjangiji.hankkimoa.place.repository.RestaurantEntity;
-import com.jjangiji.hankkimoa.place.repository.RestaurantRepository;
+import com.jjangiji.hankkimoa.restaurant.repository.RestaurantEntity;
+import com.jjangiji.hankkimoa.restaurant.repository.RestaurantRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
