+++ backend/hankkimoa/src/main/java/com/jjangiji/hankkimoa/restaurant/repository/RestaurantEntity.java
@@ -1,4 +1,4 @@
-package com.jjangiji.hankkimoa.place.repository;
+package com.jjangiji.hankkimoa.restaurant.repository;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
