+++ backend/hankkimoa/src/test/java/com/jjangiji/hankkimoa/expense/service/ExpenseServiceTest.java
@@ -4,8 +4,8 @@ import com.jjangiji.hankkimoa.IntegrationTest;
 import com.jjangiji.hankkimoa.expense.dto.ExpenseCreateRequest;
 import com.jjangiji.hankkimoa.expense.repository.ExpenseSavingGoalEntity;
 import com.jjangiji.hankkimoa.expense.repository.ExpenseSavingGoalRepository;
-import com.jjangiji.hankkimoa.place.repository.RestaurantEntity;
-import com.jjangiji.hankkimoa.place.repository.RestaurantRepository;
+import com.jjangiji.hankkimoa.restaurant.repository.RestaurantEntity;
+import com.jjangiji.hankkimoa.restaurant.repository.RestaurantRepository;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
